musicapp/server/requirements.txt
fastapi
motor
python-dotenv
